+++ Core/Mathematics/IGeoLocationFactory.cs
@@ -2,6 +2,6 @@ namespace Core.Mathematics
 {
     public interface IGeoLocationFactory
     {
-        IGeoLocation Create(decimal latitude, decimal longitude);
+        IGeoLocation Create(double latitude, double longitude);
     }
 }
